get_traj
new functions for the MD object.
md.get_traj(...) propagates the system for a given number of steps until either energy convergence or the maximal number of steps is reached. 
The intervalls at which Energy, Temperature and Frames are saved can be specified.
The location where the results will be saved can be specified aswell.